Fixed: Output in Simra orddering
VTK arrays are (to the best of my knowledge) simply row major.
So the original:

    vtkIdx = i + (j + k * ny) * nx

which corresponds to a column major array, should be:

    vtkIdx = k + (j + i * ny) * nz

The simra arrays are row major for (i,j) vs. k but column major as far
as i vs. j is concerned. In effect that only means i and j are swapped
vis-a-vis VTK ordering. So this is correct:

    simraIdx = k + (i * j * nx) * nz

Compare above and swap i <-> j and nx <-> ny.

This commit rewrites the transposition logic as an idiomatic numpy call:

    .transpose(1, 0, 2, 3)

means to keep axes 2 and 3 and just swap axes 0 and 1.